+++ tests/testthat/test-read_rmd.R
@@ -2,5 +2,5 @@ context("reading r-markdown")
 
 test_that("read_rmd is working", {
 			file <- file.path(path.package("yamlme"), "taxlistjourney.Rmd")
-			expect_is(read_rmd(file), "character")
+			expect_is(read_rmd(file), "rmd_doc")
 		})