+++ NEWS.md
@@ -8,8 +8,11 @@ yamlme 0.1.1
 
 ### Improvements
 
-* Objects of class `rmd_doc` are now structured as lists instead of character vectors.
+* Objects of class `rmd_doc` are now structured as lists instead of character
+vectors.
+* Output of function `read_rmd()` is an object of class `rmd_doc` instead of plain
+character vector.
 
 ### Bug Fixes
 
-* Dash vectors may also be of lenght 1
+* Dash vectors may also be of length 1